refactor: move admin and week pages
