+++ CoShift/frontend/src/pages/AdminPage.tsx
@@ -1,4 +1,3 @@
-// src/pages/AdminPage.tsx
 import { useState } from 'react'
 import {
     Container,
@@ -19,10 +18,10 @@ import DeleteIcon from '@mui/icons-material/Delete'
 import AddIcon from '@mui/icons-material/Add'
 import { IconButton as MuiIconButton } from '@mui/material'
 
-import AddPersonDialog from './AddPersonDialog'
-import EditPersonDialog from './EditPersonDialog'
-import { usePersons } from '../hooks/usePersons'
-import type { PersonDto } from '../../../types/person'
+import AddPersonDialog from '../features/admin/components/AddPersonDialog'
+import EditPersonDialog from '../features/admin/components/EditPersonDialog'
+import { usePersons } from '../features/admin/hooks/usePersons'
+import type { PersonDto } from '../types/person'
 
 export default function AdminPage() {
     const {
